% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlg.R
\name{graph_signal_noise_ratio}
\alias{graph_signal_noise_ratio}
\title{This function takes a list of low dimensional embedding data, and true cell type labels as input, and
compute the graph signal to noise ratio of the SNN graphs constructed with each low dimension embeddding
and the MLG graph.}
\usage{
graph_signal_noise_ratio(
  factor.list,
  knn.param = 20,
  prune.param = 1/5,
  cell_label
)
}
\arguments{
\item{factor.list}{A list that contains several low dimensional embedding data.}

\item{knn.param}{The number of neighbors used to construct k-nearest-neighbot graph}

\item{prune.param}{The prune parameter forSNN graph ??}

\item{cell_label}{True cell type label}
}
\value{
a vector of signal to noise ratio, labeled by graph name
}
\description{
This function takes a list of low dimensional embedding data, and true cell type labels as input, and
compute the graph signal to noise ratio of the SNN graphs constructed with each low dimension embeddding
and the MLG graph.
}
