% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlg.R
\name{mlg_visualization}
\alias{mlg_visualization}
\title{MLG  visualization
This function takes a list of low dimensional embedding data as input, and computes coordinates of the force-directed layout for MLG.}
\usage{
mlg_visualization(
  factor.list,
  knn.param = 20,
  prune.param = 1/5,
  label,
  label_title = "label"
)
}
\arguments{
\item{factor.list}{A list that contains several low dimensional embedding data.}

\item{knn.param}{The number of neighbors used to construct k-nearest-neighbot graph}

\item{prune.param}{The prune parameter for SNN graph. There is an edge between cell i and j in the SNN graph, if the number of common neighbors between i and j exceeds the product of knn.param and knn.param.}

\item{label}{Labels to be superimposed on the force directed layout of MLG.}
}
\value{
A ggplot object.
}
\description{
MLG  visualization
This function takes a list of low dimensional embedding data as input, and computes coordinates of the force-directed layout for MLG.
}
