% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlg.R
\name{prop.overlap.edges}
\alias{prop.overlap.edges}
\title{This function takes a list of low dimensional embedding data as input, construct a SNN graph for each low dimensional embeddings, and then plot a heatmap of the proportion of overlapping edges between each pair of SNN graphs.}
\usage{
prop.overlap.edges(factor.list, knn.param = 20, prune.param = 1/5)
}
\arguments{
\item{factor.list}{A list that contains several low dimensional embedding data.}

\item{knn.param}{The number of neighbors used to construct k-nearest-neighbot graph}

\item{prune.param}{The prune parameter forSNN graph ??}
}
\value{
A heatmap of the proportion of overlapping edges between each pair of SNN graphs constructed from low dimension embeddings.
}
\description{
This function takes a list of low dimensional embedding data as input, construct a SNN graph for each low dimensional embeddings, and then plot a heatmap of the proportion of overlapping edges between each pair of SNN graphs.
}
